+++ packages/vue-platform/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @guanxiangkai/vue-platform
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @guanxiangkai/platform-client@0.5.0
+
 ## 0.2.2
 
 ### Patch Changes